+++ modules/commons/src/include/boost_libraries_helpers.hpp
@@ -14,15 +14,15 @@
 #ifdef _MSC_VER
 #ifdef _DEBUG
 #ifdef _WIN64
-#define BOOST_TARGET "vc143-mt-gd-x64-1_85"
+#define BOOST_TARGET "vc143-mt-gd-x64-1_86"
 #else
-#define BOOST_TARGET "vc143-mt-gd-x32-1_85"
+#define BOOST_TARGET "vc143-mt-gd-x32-1_86"
 #endif
 #else
 #ifdef _WIN64
-#define BOOST_TARGET "vc143-mt-x64-1_85"
+#define BOOST_TARGET "vc143-mt-x64-1_86"
 #else
-#define BOOST_TARGET "vc143-mt-x32-1_85"
+#define BOOST_TARGET "vc143-mt-x32-1_86"
 #endif
 #endif
 #endif